+++ web/js/antragsgruen-en.js
@@ -36,7 +36,8 @@ ANTRAGSGRUEN_STRINGS = {
         "emailNumberMismatch": "The number of names and e-mail-addresses does not match.",
         "delMotionConfirm": "Do you really want to delete this motion?",
         "delAmendmentConfirm": "Do you really want to delete this amendment?",
-        "deleteMotionSectionConfirm": "Do you really want to delete this section? It will be deleted from all motions of this motion type."
+        "deleteMotionSectionConfirm": "Do you really want to delete this section? It will be deleted from all motions of this motion type.",
+        "consDeleteConfirm": "Do you really want to delete this consultation, including all motions and amendments?"
     }
 };